moved editprofile into package
