+++ app/src/main/java/com/example/neighbourproject/NeighbourApplication.kt
@@ -39,13 +39,10 @@ class NeighbourApplication: Application() {
             androidLogger()
             androidContext(this@NeighbourApplication)
 
-
             if(BuildConfig.BUILD_TYPE == "demo")
                 modules((appModuleTest)) //For test purpose
             else
                 modules((appModule)) //For a real firebase
-
-
         }
     }
 }
