+++ app/src/main/java/com/example/neighbourproject/neighbour/data/People.kt
@@ -1,8 +1,5 @@
 package com.example.neighbourproject.neighbour.data
 
-import android.text.Editable
-import android.widget.EditText
-import com.example.neighbourproject.EditProfileActivity
 import java.util.*
 
 data class People (
@@ -19,8 +16,6 @@ data class People (
         id = UUID.randomUUID().toString()
     }
 
-
-
     fun addInterest(interest: Interest) {
         interests.add(interest)
     }
@@ -28,7 +23,5 @@ data class People (
     fun removeInterest(interest: Interest) {
         interests.remove(interest)
     }
-
-
 }
