+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/ConstraineLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditProfileActivity">
+    tools:context=".ui.edit.EditProfileActivity">
 
     <ImageView
         android:id="@+id/imageView"
